Refactor configuration
