+++ src/Elastic.OpenTelemetry/Configuration/Parsers/ConfigurationParser.cs
@@ -3,11 +3,9 @@
 // See the LICENSE file in the project root for more information
 
 using System.Diagnostics.Tracing;
-using Elastic.OpenTelemetry.Configuration.Instrumentations;
 using Elastic.OpenTelemetry.Diagnostics.Logging;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using static System.StringSplitOptions;
 using static Elastic.OpenTelemetry.Configuration.Parsers.SharedParsers;
 
 namespace Elastic.OpenTelemetry.Configuration.Parsers;
@@ -31,7 +29,6 @@ internal class ConfigurationParser
 		if (string.IsNullOrEmpty(LoggingSectionLogLevel))
 			LoggingSectionLogLevel = configuration.GetValue<string>("Logging:LogLevel:Default");
 	}
-
 
 	private static void SetFromConfiguration<T>(IConfiguration configuration, ConfigCell<T> cell, Func<string, T?> parser)
 	{
@@ -75,6 +72,7 @@ internal class ConfigurationParser
 			if (sectionLogLevel < eventLogLevel)
 				eventLogLevel = sectionLogLevel;
 		}
+
 		eventLevel = eventLogLevel switch
 		{
 			LogLevel.Trace => EventLevel.Verbose,
@@ -82,87 +80,11 @@ internal class ConfigurationParser
 			LogLevel.Warning => EventLevel.Warning,
 			LogLevel.Error => EventLevel.Error,
 			LogLevel.Critical => EventLevel.Critical,
+
 			_ => EventLevel.Informational // fallback to info level
 		};
-
 	}
 
 	public void ParseSkipOtlpExporter(ConfigCell<bool?> skipOtlpExporter) =>
 		SetFromConfiguration(_configuration, skipOtlpExporter, BoolParser);
-
-	public void ParseSignals(ConfigCell<Signals?> signals) =>
-		SetFromConfiguration(_configuration, signals, SignalsParser);
-
-	public void ParseElasticDefaults(ConfigCell<ElasticDefaults?> defaults) =>
-		SetFromConfiguration(_configuration, defaults, ElasticDefaultsParser);
-
-	public void ParseInstrumentations(
-		ConfigCell<TraceInstrumentations> tracing,
-		ConfigCell<MetricInstrumentations> metrics,
-		ConfigCell<LogInstrumentations> logging
-	)
-	{
-		if (tracing.Source != ConfigSource.Environment)
-			SetFromConfiguration(_configuration, tracing, ParseTracing);
-
-		if (metrics.Source != ConfigSource.Environment)
-			SetFromConfiguration(_configuration, metrics, ParseMetrics);
-
-		if (logging.Source != ConfigSource.Environment)
-			SetFromConfiguration(_configuration, logging, ParseLogs);
-
-	}
-
-	private static IEnumerable<T>? ParseInstrumentation<T>(string? config, T[] all, Func<string, T?> getter)
-		where T : struct
-	{
-		if (string.IsNullOrWhiteSpace(config))
-			return null;
-
-		var toRemove = new HashSet<T>();
-		var toAdd = new HashSet<T>();
-
-		foreach (var token in config.Split(new[] { ';', ',' }, RemoveEmptyEntries))
-		{
-			var candidate = token.Trim();
-			var remove = candidate.StartsWith("-");
-			candidate = candidate.TrimStart('-');
-
-			var instrumentation = getter(candidate);
-			if (!instrumentation.HasValue)
-				continue;
-
-			if (remove)
-				toRemove.Add(instrumentation.Value);
-			else
-				toAdd.Add(instrumentation.Value);
-		}
-		if (toAdd.Count > 0)
-			return toAdd;
-		if (toRemove.Count > 0)
-			return all.Except(toRemove);
-		return null;
-
-	}
-
-	private static TraceInstrumentations? ParseTracing(string? tracing)
-	{
-		var instrumentations = ParseInstrumentation(tracing, TraceInstrumentationExtensions.GetValues(),
-			s => TraceInstrumentationExtensions.TryParse(s, out var instrumentation) ? instrumentation : null);
-		return instrumentations != null ? new TraceInstrumentations(instrumentations) : null;
-	}
-
-	private static MetricInstrumentations? ParseMetrics(string? metrics)
-	{
-		var instrumentations = ParseInstrumentation(metrics, MetricInstrumentationExtensions.GetValues(),
-			s => MetricInstrumentationExtensions.TryParse(s, out var instrumentation) ? instrumentation : null);
-		return instrumentations != null ? new MetricInstrumentations(instrumentations) : null;
-	}
-
-	private static LogInstrumentations? ParseLogs(string? logs)
-	{
-		var instrumentations = ParseInstrumentation(logs, LogInstrumentationExtensions.GetValues(),
-			s => LogInstrumentationExtensions.TryParse(s, out var instrumentation) ? instrumentation : null);
-		return instrumentations != null ? new LogInstrumentations(instrumentations) : null;
-	}
 }
