+++ src/Elastic.OpenTelemetry/Configuration/SdkActivationMethod.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+namespace Elastic.OpenTelemetry.Configuration;
+
+internal enum SdkActivationMethod
+{
+	NuGet,
+	AutoInstrumentation
+}
